requirements.txt
# pip install all these libraries:
pandas 
requests 
beautifulsoup4 
mysql-connector-python 
python-dotenv 
flask